chore: extract mempool polling logic into a module for sharing (#629)
