+++ src/connectors/fname/mod.rs
@@ -9,7 +9,8 @@ use tracing::{debug, error, info, warn};
 use crate::mempool::mempool::{MempoolRequest, MempoolSource};
 use crate::{
     proto::{FnameTransfer, UserNameProof, UserNameType, ValidatorMessage},
-    storage::store::{engine::MempoolMessage, node_local_state::LocalStateStore},
+    storage::store::mempool_poller::MempoolMessage,
+    storage::store::node_local_state::LocalStateStore,
     utils::statsd_wrapper::StatsdClientWrapper,
 };
