+++ src/storage/store/mod.rs
@@ -3,6 +3,7 @@ pub use self::block::*;
 pub mod account;
 pub mod block;
 pub mod engine;
+pub mod mempool_poller;
 pub mod node_local_state;
 pub mod shard;
 pub mod stores;
